revert utils.cpp/h
get color from model
