+++ src/gui/transferlistdelegate.h
@@ -31,7 +31,6 @@
 #include <QStyledItemDelegate>
 
 #include "progressbarpainter.h"
-#include "base/bittorrent/torrent.h"
 
 class TransferListDelegate final : public QStyledItemDelegate
 {
@@ -48,5 +47,4 @@ public:
 private:
     ProgressBarPainter m_progressBarPainter;
     mutable int m_nameColHeight = -1;
-    QHash<BitTorrent::TorrentState, QColor> m_stateThemeColors;
 };
