+++ src/gui/utils.cpp
@@ -40,7 +40,9 @@
 #include <QApplication>
 #include <QDesktopServices>
 #include <QPixmap>
+#include <QPixmapCache>
 #include <QPoint>
+#include <QProcess>
 #include <QRegularExpression>
 #include <QScreen>
 #include <QSize>
@@ -50,10 +52,11 @@
 #include <QWidget>
 #include <QWindow>
 
-#include "uithememanager.h"
+#include "base/global.h"
 #include "base/path.h"
 #include "base/tag.h"
-#include "base/bittorrent/torrent.h"
+#include "base/utils/fs.h"
+#include "base/utils/version.h"
 
 QPixmap Utils::Gui::scaledPixmap(const Path &path, const int height)
 {
@@ -239,43 +242,4 @@ Tag Utils::Gui::widgetTextToTag(const QString &text)
     }
 
     return Tag(cleanedText);
-}
-
-QHash<BitTorrent::TorrentState, QColor> Utils::Gui::torrentStateToColorHash()
-{
-    struct TorrentStateColorDescriptor
-    {
-        const BitTorrent::TorrentState state;
-        const QString id;
-    };
-
-    const TorrentStateColorDescriptor colorDescriptors[] =
-    {
-        {BitTorrent::TorrentState::Downloading, u"TransferList.Downloading"_s},
-        {BitTorrent::TorrentState::StalledDownloading, u"TransferList.StalledDownloading"_s},
-        {BitTorrent::TorrentState::DownloadingMetadata, u"TransferList.DownloadingMetadata"_s},
-        {BitTorrent::TorrentState::ForcedDownloadingMetadata, u"TransferList.ForcedDownloadingMetadata"_s},
-        {BitTorrent::TorrentState::ForcedDownloading, u"TransferList.ForcedDownloading"_s},
-        {BitTorrent::TorrentState::Uploading, u"TransferList.Uploading"_s},
-        {BitTorrent::TorrentState::StalledUploading, u"TransferList.StalledUploading"_s},
-        {BitTorrent::TorrentState::ForcedUploading, u"TransferList.ForcedUploading"_s},
-        {BitTorrent::TorrentState::QueuedDownloading, u"TransferList.QueuedDownloading"_s},
-        {BitTorrent::TorrentState::QueuedUploading, u"TransferList.QueuedUploading"_s},
-        {BitTorrent::TorrentState::CheckingDownloading, u"TransferList.CheckingDownloading"_s},
-        {BitTorrent::TorrentState::CheckingUploading, u"TransferList.CheckingUploading"_s},
-        {BitTorrent::TorrentState::CheckingResumeData, u"TransferList.CheckingResumeData"_s},
-        {BitTorrent::TorrentState::StoppedDownloading, u"TransferList.StoppedDownloading"_s},
-        {BitTorrent::TorrentState::StoppedUploading, u"TransferList.StoppedUploading"_s},
-        {BitTorrent::TorrentState::Moving, u"TransferList.Moving"_s},
-        {BitTorrent::TorrentState::MissingFiles, u"TransferList.MissingFiles"_s},
-        {BitTorrent::TorrentState::Error, u"TransferList.Error"_s}
-    };
-
-    QHash<BitTorrent::TorrentState, QColor> colors;
-    for (const TorrentStateColorDescriptor &colorDescriptor : colorDescriptors)
-    {
-        const QColor themeColor = UIThemeManager::instance()->getColor(colorDescriptor.id);
-        colors.insert(colorDescriptor.state, themeColor);
-    }
-    return colors;
 }
