+++ src/gui/utils.h
@@ -30,7 +30,6 @@
 #pragma once
 
 #include "base/pathfwd.h"
-#include "base/bittorrent/torrent.h"
 
 class QPixmap;
 class QPoint;
@@ -57,6 +56,4 @@ namespace Utils::Gui
 
     QString tagToWidgetText(const Tag &tag);
     Tag widgetTextToTag(const QString &text);
-
-    QHash<BitTorrent::TorrentState, QColor> torrentStateToColorHash();
 }
